requirements.txt
supabase
beautifulsoup4
requests
python-dotenv
psycopg2
